+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/missingctor/InputMissingCtor.java
@@ -1,4 +1,4 @@
-package com.puppycrawl.tools.checkstyle.checks.coding;
+package com.puppycrawl.tools.checkstyle.checks.coding.missingctor;
 
 public class InputMissingCtor
 {
